+++ tests/mem/test_fill.h
@@ -4,7 +4,7 @@
 void test_fill()
 {
     rand_seed(0);
-    for (int i = 0; i < 100; i++)
+    for (size_t i = 0; i < 100; i++)
     {
         assert_double_range(fill_rand_0_1(0), 0, 1);
         assert_double_range(fill_rand_1_1(0), -1, 1);
